+++ CHANGELOG.md
@@ -1,13 +1,17 @@
-## 0.8.5-dev1
+## 0.8.5
 
 ### Enhancements
 
+* Add parameter `skip_infer_table_types` to enable (skip) table extraction for other doc types
 * Adds optional Unstructured API unit tests in CI
 * Tracks last modified date for all document types.
 
 ### Features
 
 ### Fixes
+
+* NLTK now only gets downloaded if necessary.
+* Handling for empty tables in Word Documents and PowerPoints.
 
 ## 0.8.4